Polyfill Loan.dueDateChangedByHold to prevent errors in mylibrary
This property isn't available until mod-circulation v23.4.0, but
we were erroneously building our schema according to the latest
mod-circulation, so we allowed querying for it despite it not
returning anything.

This patches the query so that mylibrary is still allowed to do
that without erroring, until we are able to upgrade off of
mod-circulation v23.3.5, where the property doesn't exist.

Fixes https://github.com/sul-dlss/mylibrary/issues/914