Force Docker containers to use only host VM + Google for DNS
This addresses an issue reported by UIT where the container
was doing an excessive number of DNS lookups to UIT servers.

The container ordinarily has the same /etc/resolv.conf as its
host, but in this case that's undesirable because it allows
the container to discover UIT DNS and make lots of queries
without caching.

Instead, we want to force the container to point at its host's
DNS, which does use dnsmasq and will avoid so many non-cached
lookups.

We keep Google's DNS in as a backup because they can handle it.